backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
openai==1.10.0
weaviate-client==4.4.0
pytesseract==0.3.10
pdf2image==1.17.0
pymupdf==1.23.21
python-multipart==0.0.6
transformers==4.37.2
torch==2.2.0
pydantic==2.6.0
pydantic-settings==2.1.0
